% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capper.R
\name{capper}
\alias{capper}
\title{Capper}
\usage{
capper(x, floor, ceiling)
}
\arguments{
\item{x}{number or vector to cap}

\item{floor}{floor of cap}

\item{ceiling}{ceiling of cap}
}
\value{
a number either equal to x, floor or ceiling, depending on if x is
smaller than floor or larger than ceiling, or between floor and ceiling.
}
\description{
a function that cuts a number between a floor and a ceiling.
}
