% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct-andre.R
\name{correct_andre}
\alias{correct_andre}
\title{Change Andre or Andrè to André}
\usage{
correct_andre(x)
}
\arguments{
\item{x}{Input vector. Either a character vector, or something coercible to one.}
}
\value{
A character vector.
}
\description{
Change Andre or Andrè to André
}
