% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discount_cashflow.R
\name{discount_cashflow}
\alias{discount_cashflow}
\title{Discount Cashflow}
\usage{
discount_cashflow(df, cf, rate, period)
}
\arguments{
\item{df}{Data frame with cash flow, interest rate and period}

\item{cf}{Cash flow}

\item{rate}{Interest rate}

\item{period}{Period, e.g. year.}
}
\value{

}
\description{
This function discounts a cash flow with the interest and return the net present value (NPV). The function assumes
that the first period of \code{df} is period zero in the npv calculation, thus the cash flows in this period is not
discounted.
}
\examples{

df <- tibble(
cf = c(100, 200, 300),
rate = c(0.05, 0.10, 0.20),
period = c(2020, 2021, 2022)
)

df_npv <- discount_cashflow(df, cf, rate, period)


}
