% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psum.R
\name{psum}
\alias{psum}
\title{psum}
\usage{
psum(...)
}
\arguments{
\item{...}{list of vectors of equal dimensions}
}
\value{
a vector of equal dimensions as the input vector with the rowsum of input vectors
}
\description{
function that does rowsums of vectors
}
