% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rolling_aggregate.r
\name{rolling_aggregate}
\alias{rolling_aggregate}
\title{Add count and aggregated column to data.table}
\usage{
rolling_aggregate(
  df,
  date,
  window_size,
  input_col,
  output_name,
  aggregate_func,
  ...
)
}
\arguments{
\item{df}{data.table}

\item{date}{Column of format date}

\item{window_size}{Size of prior window in days}

\item{input_col}{Input column to aggregate function}

\item{output_name}{Name of aggregated column}

\item{aggregate_func}{Aggregate function to be applied to input_col}

\item{...}{Arguments to aggregate function}
}
\value{
Original data.table with count and aggregated columns
}
\description{
Add count and aggregated column to data.table
}
