% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_safely.R
\name{safely_map_dfr}
\alias{safely_map_dfr}
\title{Apply a function to each element of a vector safely}
\usage{
safely_map_dfr(.x, .f, ..., .id = NULL)
}
\arguments{
\item{.x}{A list or atomic vector.}

\item{.f}{A function, formula, or vector (not necessarily atomic).

If a \strong{function}, it is used as is.

If a \strong{formula}, e.g. \code{~ .x + 2}, it is converted to a function. There
are three ways to refer to the arguments:
\itemize{
\item For a single argument function, use \code{.}
\item For a two argument function, use \code{.x} and \code{.y}
\item For more arguments, use \code{..1}, \code{..2}, \code{..3} etc
}

This syntax allows you to create very compact anonymous functions.

If \strong{character vector}, \strong{numeric vector}, or \strong{list}, it is
converted to an extractor function. Character vectors index by
name and numeric vectors index by position; use a list to index
by position and name at different levels. If a component is not
present, the value of \code{.default} will be returned.}

\item{...}{Additional arguments passed on to the mapped function.}

\item{.id}{Either a string or \code{NULL}. If a string, the output will contain
a variable with that name, storing either the name (if \code{.x} is named) or
the index (if \code{.x} is unnamed) of the input. If \code{NULL}, the default, no
variable will be created.

Only applies to \code{_dfr} variant.}
}
\value{
Funksjonen returnerer et listeobjekt med to elementer.
$result inneholder en data.frame med resultat for de elementene som kjørte suksessfullt,
mens $error er en liste med samme lengde som .x. Elementene i listen er NULL for de elementene i .x
som kjørte uten feil og feilobjektet for de som feilet.
}
\description{
Denne funksjonen er en wrapper rundt funksjonene map_dfr og safely fra purrr.
}
\details{
Se hjelpeside for purrr::map for doks.
}
\examples{
a <- safely_map_dfr(c(-1, 1), ~ data.frame(a = rnorm(.)))
}
