% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magimisc-package.R
\docType{package}
\name{magimisc-package}
\alias{magimisc}
\alias{magimisc-package}
\title{magimisc: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Test Testson \email{test.testson@test.com}

}
\keyword{internal}
