% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabset_list.R
\name{tabset_list}
\alias{tabset_list}
\title{Create tabset from a list}
\usage{
tabset_list(list, header_level)
}
\arguments{
\item{list}{A \code{list} of elements to be printed in each tab}

\item{header_level}{An \code{integer} specifying the level of the headers}
}
\value{

}
\description{
Create tabset from a list
}
\examples{


}
